troca de provider para armazenamento de fotos

+++ src/modules/eventos/services/UpdateEventoService.ts
@@ -28,36 +28,43 @@ export class UpdateEventoService {
     private cacheService: ICacheService,
     @inject('IStorageService')
     private storageService: IStorageService
-  ) {}
-
+  ) {
+    this.cacheService = cacheService;
+    this.eventoRepository = eventoRepository;
+    this.photoRepository = photoRepository
+  }
   public async execute(id: string, evento: IRequest): Promise<IEvento> {
     const eventoExists = await this.eventoRepository.findById(id);
     if (!eventoExists) {
-      throw new AppError('Evento não encontrado', 404);
+      throw new AppError('Evento não encontrado', 404)
     }
 
-    const photos = await this.photoRepository.findAllPhotosByEventId(eventoExists._id);
-    if (photos.length < 1) throw new AppError('nenhuma foto existente', 404);
+    //APAGANDO FOTOS DO STORAGE
+    const photos = await this.photoRepository.findAllPhotosByEventId(eventoExists._id)
+    if (photos.length < 1) throw new AppError('nenhuma foto existente', 404)
 
-    // APAGANDO FOTOS DO STORAGE e atualizando URLs
     await Promise.all(photos.map(async (photo) => {
-      await this.storageService.deleteFile(photo.url);
-      photo.url = evento.photos.shift() as string; // Atualiza com a nova foto
-      if (photo.url) {
-        await this.storageService.saveFile(photo.url);
+      const file =this.storageService.getFile(photo.url)
+      if (file) {
+        await this.storageService.deleteFile(photo.url)
       }
-      await this.photoRepository.update(photo._id, photo);
-    }));
 
-    const updatedEvent = await this.eventoRepository.update(id, {
+      evento.photos.map(async (url) => {
+        photo.url = evento.photos.shift() as string
+        await this.storageService.saveFile(url)
+        console.log('foto salva:', photo.url)
+      })
+      await this.photoRepository.update(photo._id, photo)
+    }))
+   const event =  await this.eventoRepository.update(id, {
       address: evento.address,
       data_fim: evento.data_fim,
       data_inicio: evento.data_inicio,
       descricao: evento.descricao,
       titulo: evento.titulo
-    } as IUpdateEvento);
-
-    await this.cacheService.invalidate('api_anjobom_EVENTS_LIST');
-    return updatedEvent as IEvento;
+    }as IUpdateEvento)
+    await this.cacheService.invalidate('api_anjobom_EVENTS_LIST')
+    return event as IEvento;
   }
+
 }
